cleaned up style and code
