--- farm-animal/src/styles/Info.css
@@ -1,131 +1,0 @@
-*, *::before, *::after{
-    box-sizing: inherit;
-}
-
-h1{
-    text-align: center;
-}
-
-h2 {
-    text-align: center;
-}
-
-p {
-    text-align: justify;
-    font-size: x-large;
-}
-
-.about-section{
-    padding: 50px;
-    text-align: center;
-    background-color: var(--primary);
-    color:black;
-    border-radius: 25px;
-}
-
-.info-column{
-    float: left;
-    /* width: 100%; */
-    margin-bottom: 2%;
-    margin-top: 2%;
-    padding: 0 8px;
-}
-
-.info-card{
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-    margin: 2%;
-    background-color: #fec89a;
-    border-radius: 25px;
-}
-
-.info-container::after, .row::after{
-    content: "";
-    clear: both;
-    /* display: flex;
-    align-items: center;
-    flex-direction: column; */
-}
-
-.info-container{
-    word-wrap: break-word;
-}
-
-.entire-page{
-    background-color: #fff1f0;
-    height: 100%;
-    padding: 5%;
-}
-
-.info-about{
-    margin-left: 25%;
-    margin-right: 25%;
-}
-
-
-.info-row{
-    display: flex;
-    /* align-items: center; */
-    align-content: space-evenly;
-    text-align: center;
-    justify-content: center;
-}
-
-.meet-the-team{
-    padding-top: 2%;
-    padding-bottom: 2%;
-}
-
-.card-title{
-    color: grey;
-    text-align: center;
-}
-
-.card-body{
-    text-align: center;
-}
-
-.card-email{
-    text-align: center;
-}
-@media screen and (max-width: 650px) {
-    .column {
-      width: 100%;
-      display: block;
-    }
-  }
-
-.info-game-detail-section{
-    display:flex;
-    margin-bottom: 2%;
-    margin-top: 2%;
-}
-
-.info-how-to-play{
-    width: 40%;
-    float: left;
-}
-
-.info-about-mafia{
-    width: 60%;
-    float: right;
-}
-
-.info-GDS-colomn{
-    padding: 2%;
-    font-size: x-large;
-}
-
-.info-GDS-card{
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-    margin: 2%;
-    background-color: #fec89a;
-    padding-left: 1%;
-    padding-right: 1%;
-    border-radius: 25px;
-    text-align: justify;
-}
-
-.info-GDS-header{
-    text-align: center;
-    padding-top: 1%;
-}
